requirements.txt
# Core
pyyaml>=6.0
pydantic>=2.0
jsonschema>=4.0

# Data processing
numpy>=1.24
pandas>=2.0
pyarrow>=14.0

# CLI
click>=8.0
rich>=13.0

# Testing
pytest>=7.0
pytest-cov>=4.0

# Infrastructure (optional, for ClickHouse loading)
clickhouse-connect>=0.7.0

# Dev
black>=23.0
ruff>=0.1.0
mypy>=1.0
